ImageOnMap is now a zLib plugin.
* NEW: added zLib dependency.
* NEW: removed all classes moved to the zLib, and updated the references to these classes to the zLib ones.
* BUG: fixed some GUI navigation problems.
* NEW: implemented the items renaming into the details GUI.
