+++ src/main/java/fr/moribus/imageonmap/commands/maptool/MigrateCommand.java
@@ -18,21 +18,18 @@
 
 package fr.moribus.imageonmap.commands.maptool;
 
-import fr.moribus.imageonmap.commands.*;
+import fr.moribus.imageonmap.commands.IoMCommand;
 import fr.moribus.imageonmap.migration.MigratorExecutor;
+import fr.zcraft.zlib.components.commands.CommandException;
+import fr.zcraft.zlib.components.commands.CommandInfo;
 import org.bukkit.command.CommandSender;
 
-@CommandInfo(name = "migrate")
-public class MigrateCommand extends Command
+@CommandInfo (name = "migrate")
+public class MigrateCommand extends IoMCommand
 {
-    public MigrateCommand(Commands commandGroup) {
-        super(commandGroup);
-    }
-    
     @Override
     protected void run() throws CommandException
     {
-        final CommandSender cmdSender = sender;
         if(MigratorExecutor.isRunning())
         {
             error("A migration process is already running. Check console for details.");
